Add branch version to helm chart
When building this chart, it will default to using the version specified on line 20.
If using `helm package`, even with the version specified using the `--app-version` flag, the chart will still be recognized by the default Version value eg. when going to use `helm push`.
If taring the directory and naming the version with the tar command, again `helm push` only notices the metadata from this file. This results in errors when pushing to NGC.

Error message:
```
helm push dicom-adapter-0.8.1-rc1.tgz ngc
Pushing dicom-adapter-0.0.0.0.tgz to ngc...
Error: 409: nvidia/clara/dicom-adapter-0.0.0.0.tgz already exists
```

This value needs to be incremented in tandem with the Version file in the root directory (and for release branches).